removed the experiment class
