+++ templates/decode_report.html
@@ -80,7 +80,7 @@
 <body>
     <div class="container">
         <h1>DELi Decode Report</h1>
-        <h4>Report generated on {{ timestamp }} for experiment {{ experiment }} and selection {{ selection }}</h4>
+        <h4>Report generated on {{ timestamp }} for selection {{ selection }}</h4>
         <p>Selection run on: {{ run_date }}</p>
         <p>Target: {{ target }}</p>
         <p>Selection condition: {{ selection_cond }}</p>
